Move out ImmutableArrayBuilderDebuggerProxy
