+++ src/System.Collections.Immutable/src/System.Collections.Immutable.csproj
@@ -38,6 +38,7 @@
     <Compile Include="System\Collections\Immutable\IOrderedCollection.cs" />
     <Compile Include="System\Collections\Immutable\ImmutableArray.cs" />
     <Compile Include="System\Collections\Immutable\ImmutableArray_1.Builder.cs" />
+    <Compile Include="System\Collections\Immutable\ImmutableArray_1.Builder.DebuggerProxy.cs" />
     <Compile Include="System\Collections\Immutable\ImmutableArray_1.Enumerator.cs" />
     <Compile Include="System\Collections\Immutable\ImmutableArray_1.cs" />
     <Compile Include="System\Collections\Immutable\ImmutableDictionary.cs" />
