+++ src/System.Collections.Immutable/src/System/Collections/Immutable/ImmutableArray_1.Builder.cs
@@ -4,7 +4,6 @@
 
 using System.Collections.Generic;
 using System.Diagnostics;
-using System.Diagnostics.CodeAnalysis;
 using System.Diagnostics.Contracts;
 
 namespace System.Collections.Immutable
@@ -738,37 +737,6 @@ namespace System.Collections.Immutable
                 {
                     nodes[offset + i] = items[i];
                 }
-            }
-        }
-    }
-
-    /// <summary>
-    /// A simple view of the immutable collection that the debugger can show to the developer.
-    /// </summary>
-    internal sealed class ImmutableArrayBuilderDebuggerProxy<T>
-    {
-        /// <summary>
-        /// The collection to be enumerated.
-        /// </summary>
-        private readonly ImmutableArray<T>.Builder _builder;
-        /// <summary>
-        /// Initializes a new instance of the <see cref="ImmutableArrayBuilderDebuggerProxy{T}"/> class.
-        /// </summary>
-        /// <param name="builder">The collection to display in the debugger</param>
-        public ImmutableArrayBuilderDebuggerProxy(ImmutableArray<T>.Builder builder)
-        {
-            _builder = builder;
-        }
-
-        /// <summary>
-        /// Gets a simple debugger-viewable collection.
-        /// </summary>
-        [DebuggerBrowsable(DebuggerBrowsableState.RootHidden)]
-        public T[] A
-        {
-            get
-            {
-                return _builder.ToArray();
             }
         }
     }
